[#145]refect:랜딩페이지 시작하기섹션 분리
